prediction-microservice/requirements.txt
fastapi==0.109.2
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.5.0
tensorflow==2.15.0
tensorflow_intel==2.15.0
python-multipart